+++ swift/Clibautograph.xcframework/ios-arm64/Headers/autograph/sign.h
@@ -8,7 +8,7 @@ extern "C" {
 int autograph_sign_subject(unsigned char *signature,
                            const unsigned char *private_key,
                            const unsigned char *subject,
-                           const unsigned long long subject_size);
+                           const unsigned int subject_size);
 
 #ifdef __cplusplus
 }  // extern "C"
